Fix commit gates: move tests to separate file and add lint attributes
- Move action_payloads tests to separate action_payloads_tests.rs file
  to resolve module-max-lines lint error
- Add expect attributes for clippy::unwrap_used and clippy::float_cmp
  in test file with appropriate reasons
- Fix doc markdown lint by adding backticks
